# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Airport API",
        "description": "API documentation for Airport search service",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/airports/searchByCode": {
            "get": {
                "tags": [
                    "Code"
                ],
                "summary": "Search by code",
                "description": "Returns a list of airports that match the provided code",
                "operationId": "1ef6dba060c60aaaabaea5970fef6de2",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Airport"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No cities found"
                    }
                }
            }
        },
        "/api/airports/searchByCity": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Search cities by name",
                "description": "Returns a list of cities that match the provided query",
                "operationId": "297477673e449bd396ca2eb38232ce55",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Airport"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No airports found"
                    }
                }
            }
        },
        "/api/airports/searchByAirport": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Search airports by name",
                "description": "Returns a list of airports that match the provided query",
                "operationId": "75a793469aac1505b44b94cca6849dc4",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Airport"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No airports found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Airport": {
                "title": "Airport",
                "description": "Airport data",
                "properties": {
                    "code": {
                        "description": "The airport code",
                        "type": "string"
                    },
                    "city_name_ru": {
                        "description": "The city name in Russian",
                        "type": "string"
                    },
                    "city_name_en": {
                        "description": "The city name in English",
                        "type": "string"
                    },
                    "airport_name_ru": {
                        "description": "The airport name in Russian",
                        "type": "string"
                    },
                    "airport_name_en": {
                        "description": "The airport name in English",
                        "type": "string"
                    },
                    "country": {
                        "description": "The country code",
                        "type": "string"
                    },
                    "latitude": {
                        "description": "The latitude coordinate of the airport",
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "description": "The longitude coordinate of the airport",
                        "type": "number",
                        "format": "float"
                    },
                    "timezone": {
                        "description": "The timezone of the airport",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}